requirements/prod.txt
Django==5.1.5
python-dotenv==1.0.1
djangorestframework==3.15.2
parameterized==0.9.0
